add load test, fix references issues

+++ src/Solution/Demos/Reusables/Repositories/IProductsRepository.cs
@@ -7,11 +7,11 @@ namespace Reusables.Repositories
 {
     public interface IProductsRepository
     {
-        Task<Product> GetProduct(long productId, CancellationToken cancellationToken = default);
-        Task<Product> GetProductFull(long productId, CancellationToken cancellationToken = default);
+        Task<Product> GetProduct(int productId, CancellationToken cancellationToken = default);
+        Task<Product> GetProductFull(int productId, CancellationToken cancellationToken = default);
         Task<List<Product>> GetProductsPage(int page = 1, int pageSize = 20, CancellationToken cancellationToken = default);
         Task<List<Product>> GetProductsPageFull(int page = 1, int pageSize = 20, CancellationToken cancellationToken = default);
         Task<long> GetTotalProducts(CancellationToken cancellationToken = default);
-        Task EditProductName(long productId, string productName);
+        Task EditProductName(int productId, string productName);
     }
 }
